+++ src/pages/element/index.js
@@ -3,7 +3,7 @@ import React from "react"
 import { Link, StaticQuery, graphql } from "gatsby"
 import Layout from "../../components/layout"
 import SEO from "../../components/seo"
-import Button from "../../components/filtering"
+import Button from "../../components/button"
 
 class ElementList extends React.Component {
   constructor() {
